+++ src/main/java/zhibi/admin/role/domain/RoleMenu.java
@@ -2,7 +2,8 @@ package zhibi.admin.role.domain;
 
 import lombok.Data;
 import lombok.experimental.Accessors;
-import zhibi.admin.role.common.base.dto.BaseDomain;
+import zhibi.fast.mybatis.dto.BaseDomain;
+
 
 import javax.persistence.Column;
 import javax.persistence.Table;
@@ -15,9 +16,9 @@ import javax.persistence.Table;
 @Accessors(chain = true)
 public class RoleMenu extends BaseDomain {
     @Column(name = "role_id")
-    private Integer roleId;
+    private Long roleId;
 
     @Column(name = "menu_id")
-    private Integer menuId;
+    private Long menuId;
 
 }
