+++ src/main/java/zhibi/admin/role/domain/UserRole.java
@@ -2,7 +2,8 @@ package zhibi.admin.role.domain;
 
 import lombok.Data;
 import lombok.experimental.Accessors;
-import zhibi.admin.role.common.base.dto.BaseDomain;
+import zhibi.fast.mybatis.dto.BaseDomain;
+
 
 import javax.persistence.Column;
 import javax.persistence.Table;
@@ -15,8 +16,8 @@ import javax.persistence.Table;
 @Accessors(chain = true)
 public class UserRole extends BaseDomain {
     @Column(name = "user_id")
-    private Integer userId;
+    private Long userId;
 
     @Column(name = "role_id")
-    private Integer roleId;
+    private Long roleId;
 }
